Comment 사용자가 수정 삭제 위해 입력한 비밀번호와 DB 저장 비밀번호 일치 확인
- commentId 해당하는 레코드 존재 여부 먼저 확인
  - 각 경우에 따라 다른 방식으로 예외처리 하기 위함
- 존재하는 경우 비밀번호 검사 private 메서드 사용
  - 암호화 될 password는 서버까지 끌고 오지 않고, DB 서버 내에서 비교하도록 할 것

+++ week4/src/main/kotlin/kotlinassignment/week4/domain/comment/service/CommentService.kt
@@ -4,6 +4,7 @@ import kotlinassignment.week4.domain.comment.dto.CommentCreateRequest
 import kotlinassignment.week4.domain.comment.dto.CommentDeleteRequest
 import kotlinassignment.week4.domain.comment.dto.CommentResponse
 import kotlinassignment.week4.domain.comment.dto.CommentUpdateRequest
+import kotlinassignment.week4.domain.comment.model.Comment
 import kotlinassignment.week4.domain.comment.model.fromRequestToComment
 import kotlinassignment.week4.domain.comment.model.toResponse
 import kotlinassignment.week4.domain.comment.model.updateFrom
@@ -36,8 +37,11 @@ class CommentService(
 
     @Transactional
     fun updateComment(toDoCardId: Long, commentId: Long, request: CommentUpdateRequest): CommentResponse {
-        val targetComment =
-            commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("Comment", commentId)
+        // 해당 commentId 존재 여부만 먼저 확인 후, 존재하지 않는 경우 예외 처리, 존재하는 경우 비밀번호 검사 후 entity 가져옴
+        // - select를 두 번 이상 하게 되나, 각 경우에 따라 다른 방식으로 예외처리하기 위해 이렇게 구현하고자 했음
+        if (!commentRepository.existsById(commentId)) throw ModelNotFoundException("Comment", commentId)
+        val targetComment = getCommentIfPasswordMatches(commentId, request.password)
+
         if (targetComment.toDoCard.id != toDoCardId) {
             TODO("새로운 에러 클래스 만들어서 던지게 하기 - path variable로 받은 toDoCardId와 Comment가 참조하는 ToDoCard의 id가 일치하지 않는 경우")
         }
@@ -50,12 +54,21 @@ class CommentService(
 
     @Transactional
     fun deleteComment(toDoCardId: Long, commentId: Long, request: CommentDeleteRequest): Unit {
-        val targetComment =
-            commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("Comment", commentId)
+        // 해당 commentId 존재 여부만 먼저 확인 후, 존재하지 않는 경우 예외 처리, 존재하는 경우 비밀번호 검사 후 entity 가져옴
+        // - select를 두 번 이상 하게 되나, 각 경우에 따라 다른 방식으로 예외처리하기 위해 이렇게 구현하고자 했음
+        if (!commentRepository.existsById(commentId)) throw ModelNotFoundException("Comment", commentId)
+        val targetComment = getCommentIfPasswordMatches(commentId, request.password)
+
         if (targetComment.toDoCard.id != toDoCardId) {
             TODO("새로운 에러 클래스 만들어서 던지게 하기 - path variable로 받은 toDoCardId와 Comment가 참조하는 ToDoCard의 id가 일치하지 않는 경우")
         }
 
         commentRepository.deleteById(commentId)
+    }
+
+    private fun getCommentIfPasswordMatches(commentId: Long, passwordRequested: String): Comment { // readOnly = true, propagation = PROPAGATION.REQUIRED
+        // 암호화 될 password는 서버까지 끌고 오지 않고, DB 서버 내에서 비교 후 일치하면 처리되도록 할 것 
+        return commentRepository.findByIdAndPassword(commentId, passwordRequested)
+            ?: TODO("새로운 에러 클래스 만들어서 던지게 하기 - Comment 비밀번호가 일치하지 않는 경우(DB의 record의 password와 request의 password)")
     }
 }
